gas_counter moved to mod
