+++ runtime/near-vm-logic/src/lib.rs
@@ -2,6 +2,7 @@ mod config;
 mod context;
 mod dependencies;
 mod logic;
+mod gas_counter;
 #[cfg(feature = "mocks")]
 pub mod mocks;
 pub mod serde_with;
